➕ add security check to all the user based api through the use of a "api_key"

+++ src/routes/api/user/[id].ts
@@ -2,87 +2,141 @@ import supabaseClient from "$lib/utils/supabaseClient"
 import type { RequestHandlerOutput } from "@sveltejs/kit"
 import type { RequestEvent } from "@sveltejs/kit/types/internal"
 
-export async function post({ params, request }:RequestEvent): Promise<RequestHandlerOutput> {
-    let id = params.id
-    let body = await request.json()
-    let hasError = false
-    let errorMessageList = Array<string>()
+const {
+    VITE_API_KEY
+} = import.meta.env
 
-    body.user_metadata = {}
+export async function post({ params, request, url }:RequestEvent): Promise<RequestHandlerOutput> {
+    let apiKey = url.searchParams.get("api_key")
 
-    let idChecker = await supabaseClient
-        .from("users")
-        .select("*")
-        .eq("id", id)
-        .limit(1)
-        .single()
-    if (idChecker.data) {
-        if (body.username) {
-            let { data } = await supabaseClient
-                .from("users")
-                .select("id")
-                .eq("username", body.username)
-                .limit(1)
-                .single() 
-            if (data) {
-                if (data.id != id) {
+    if (apiKey == VITE_API_KEY) {
+        let id = params.id
+        let body = await request.json()
+        let hasError = false
+        let errorMessageList = Array<string>()
+
+        body.user_metadata = {}
+
+        let idChecker = await supabaseClient
+            .from("users")
+            .select("*")
+            .eq("id", id)
+            .limit(1)
+            .single()
+        if (idChecker.data) {
+            if (body.username) {
+                let { data } = await supabaseClient
+                    .from("users")
+                    .select("id")
+                    .eq("username", body.username)
+                    .limit(1)
+                    .single() 
+                if (data) {
+                    if (data.id != id) {
+                        delete body.username
+                        hasError = true
+                        errorMessageList.push("Username already exist")
+                    }
+                } else {
+                    body.user_metadata.username = body.username
                     delete body.username
-                    hasError = true
-                    errorMessageList.push("Username already exist")
                 }
-            } else {
-                body.user_metadata.username = body.username
-                delete body.username
             }
-        }
-    
-        if (body.email) {
-            let { data } = await supabaseClient
-                .from("users")
-                .select("id")
-                .eq("email", body.email)
-                .limit(1)
-                .single() 
-            if (data) {
-                if (data.id != id) {
-                    delete body.email
-                    hasError = true
-                    errorMessageList.push("Email already exist")
+        
+            if (body.email) {
+                let { data } = await supabaseClient
+                    .from("users")
+                    .select("id")
+                    .eq("email", body.email)
+                    .limit(1)
+                    .single() 
+                if (data) {
+                    if (data.id != id) {
+                        delete body.email
+                        hasError = true
+                        errorMessageList.push("Email already exist")
+                    } 
+                } else {
+                    body.user_metadata.email_confirm = false
+                }
+            }
+        
+            if (body.password) {     
+                body.user_metadata.password = body.password
+            }
+        
+            let { data } = await supabaseClient.auth.api.updateUserById( id, body )
+        
+            if (hasError) {
+                return {
+                    status: 400,
+                    body: {
+                        code: 400,
+                        message: errorMessageList
+                    }
+                }
+            } else if (data) {
+                return {
+                    status: 200,
+                    body: {
+                        email: data.email,
+                        password: data.user_metadata.password,
+                        username: data.user_metadata.username
+                    }
                 } 
             } else {
-                body.user_metadata.email_confirm = false
+                return {
+                    status: 400,
+                    body: {
+                        code: 400,
+                        message: "Backend error"
+                    }
+                }
             }
-        }
-    
-        if (body.password) {     
-            body.user_metadata.password = body.password
-        }
-    
-        let { data } = await supabaseClient.auth.api.updateUserById( id, body )
-    
-        if (hasError) {
+        } else {
             return {
                 status: 400,
                 body: {
                     code: 400,
-                    message: errorMessageList
+                    message: "User not found"
                 }
             }
-        } else if (data) {
+        }
+    } else {
+        return {
+            status: 400,
+            body: {
+                code: 400,
+                message: "Incorrect api key"
+            }
+        }
+    }
+    
+    
+}
+
+export async function get({ params, url }:RequestEvent): Promise<RequestHandlerOutput> {
+    let apiKey = url.searchParams.get("api_key")
+
+    if (apiKey == VITE_API_KEY) {
+        let id = params.id
+        let { data } = await supabaseClient
+            .from("users")
+            .select("*")
+            .eq("id", id)
+            .limit(1)
+            .single()
+        if (data) {
             return {
                 status: 200,
-                body: {
-                    email: data.email,
-                    password: data.user_metadata.password,
-                    username: data.user_metadata.username
-                }
-            } 
+                body: data
+            }
         } else {
             return {
                 status: 400,
                 body: {
                     code: 400,
-                    message: "Backend error"
+                    message: "User not found"
                 }
             }
         }
@@ -91,57 +145,45 @@ export async function post({ params, request }:RequestEvent): Promise<RequestHan
             status: 400,
             body: {
                 code: 400,
-                message: "User not found"
+                message: "Incorrect api key"
             }
         }
     }
     
 }
 
-export async function get({ params }:RequestEvent): Promise<RequestHandlerOutput> {
-    let id = params.id
-    let { data } = await supabaseClient
-        .from("users")
-        .select("*")
-        .eq("id", id)
-        .limit(1)
-        .single()
-    if (data) {
-        return {
-            status: 200,
-            body: data
-        }
-    } else {
-        return {
-            status: 400,
-            body: {
-                code: 400,
-                message: "User not found"
-            }
-        }
-    }
-}
+export async function del({ params, url }:RequestEvent): Promise<RequestHandlerOutput> {
+    let apiKey = url.searchParams.get("api_key")
 
-export async function del({ params }:RequestEvent): Promise<RequestHandlerOutput> {
-    let id = params.id
-    let idChecker = await supabaseClient
-        .from("users")
-        .select("*")
-        .eq("id", id)
-        .limit(1)
-        .single()
-    if (idChecker.data)  {
-        let { error } = await supabaseClient.auth.api.deleteUser(id) 
-        if (!error) {
-            return {
-                status: 200
+    if (apiKey == VITE_API_KEY) {
+        let id = params.id
+        let idChecker = await supabaseClient
+            .from("users")
+            .select("*")
+            .eq("id", id)
+            .limit(1)
+            .single()
+        if (idChecker.data)  {
+            let { error } = await supabaseClient.auth.api.deleteUser(id) 
+            if (!error) {
+                return {
+                    status: 200
+                }
+            } else {
+                return {
+                    status: 400,
+                    body: {
+                        code: 400,
+                        message: "Backend error"
+                    }
+                }
             }
         } else {
             return {
                 status: 400,
                 body: {
                     code: 400,
-                    message: "Backend error"
+                    message: "User not found"
                 }
             }
         }
@@ -150,8 +192,10 @@ export async function del({ params }:RequestEvent): Promise<RequestHandlerOutput
             status: 400,
             body: {
                 code: 400,
-                message: "User not found"
+                message: "Incorrect api key"
             }
         }
     }
+
+    
 }
